SBRFNDFL-8729 - Не удалось создать форму "6-НДФЛ" Проверил под отладкой на unstable

+++ src/main/resources/form_template/ndfl/report_6ndfl/v2016/report_6ndfl.groovy
@@ -788,7 +788,7 @@ class Report6Ndfl extends AbstractScriptClass {
                     ? 0
                     : incomeList.findAll { income -> isBelongToJanuary2019Period(income.taxDate) && income.refoundTax }
                     .collect { income -> income.refoundTax}
-                    .sum()
+                    .sum() as Long
         }
 
         void buildRows(IncomeList incomeList) {
